backend/requirements.txt
fastapi
uvicorn[standard]
aiosqlite
python-dotenv
groq
httpx
pydantic